// lib/components/Sidebar/Sidebar.variants.d.ts
export declare const wrapperSiderbarVariants: (props?: ({
    mode?: "expanded" | "collapsed" | null | undefined;
} & import('class-variance-authority/types').ClassProp) | undefined) => string;
export declare const dragVariants: (props?: import('class-variance-authority/types').ClassProp | undefined) => string;